requirements.txt
opencv-python
Pillow
numpy
rtsp